# More lenient base64 decoder, base64 performance improvements