--- a/config/config-options/DUK_USE_BASE64_SUPPORT.yaml
+++ b/config/config-options/DUK_USE_BASE64_SUPPORT.yaml
@@ -1,0 +1,7 @@
+define: DUK_USE_BASE64_SUPPORT
+introduced: 2.3.0
+default: true
+tags:
+  - codec
+description: >
+  Enable base64 encoding/decoding support.