--- a/config/tags.yaml
+++ b/config/tags.yaml
@@ -89,6 +89,9 @@ torture:
     Development time options to stress test corner case handling by e.g.
     causing a garbage collection on every allocation.
 
+codec:
+  title: Codecs
+
 encoding-api:
   title: WHATWG Encoding API