… src-input/duk_api_codec.c
… src-input/duk_util_misc.c
… tests/ecmascript/test-base64-dec-brute.js
… tests/ecmascript/test-base64-enc-brute.js
… tests/ecmascript/test-dev-base64-enc.js