- [x] Allow missing ('Zm'), partial ('Zm='), and extra padding ('Zm====') in the base64 decoder
- [x] Improve base-64 fast path encoder and decoder performance
- [x] Improve base-64 slow path encoder and decoder performance without increasing footprint
- [x] Base64 testcase coverage improvements
- [x] Make base64 support optional (DUK_USE_BASE64_SUPPORT) and disable it by default in lowmem configuration
- [x] Make hex support optional (DUK_USE_HEX_SUPPORT) and disable it by default in lowmem configuration
- [x] 2.3 migration note for base64/hex and lowmem
- [x] Releases entry

Performance improvements for fast path on my development laptop (Intel(R) Core(TM) i7-4600U CPU @ 2.10GHz) measured using tests/perf/test-base64* (includes string intern after encode):
- Base64 encode: 1030 MB/sec in master, 1170 MB/sec after
- Base64 decode: 1770 MB/sec in master, 1900 MB/sec after

Performance improvements for slow path:
- Base64 encode: 130 MB/sec in master, 360 MB/sec after
- Base64 decode: 440 MB/sec in master, 450 MB/sec after (negligible)