@svaarala 2017-10-30
There is considerable variation in base64 decoder leniency. After the changes in this pull, the remaining differences to Node.js behavior is:
- Node.js doesn't accept concatenated documents but terminates at the first padding character, so `Zm==Zm9v` decodes to `f`, whereas Duktape decodes to `ffoo`.
- Duktape rejects `Z`, `Z===`, and other padding variants of the form with a TypeError, Node.js accepts them and decodes an empty string because there's only 6 bits of input.