@fatcerberus 2017-10-30
I'm curious what the difference is (algorithm-wise) between the fast and slow paths for base64 and what the conditions for using each is, as base64 is a very simple encoding--or at least seems to be.