@svaarala 2017-10-30
My wording was a bit imprecise, what I mean is the fast variant (= DUK_USE_BASE64_FASTPATH enabled) and slow variant (disabled). The fast variant has a fast path for dense groups (`"Zm9v"`), in fact for two dense groups at a time, and other kinds of groups that involve whitespace (`"Zm\n9v"`), padding, etc.

For decoding base-64 is definitely not a simple format due to all the variation, see e.g. https://en.wikipedia.org/wiki/Base64#Variants_summary_table. Unfortunately even with the leniency changes here, the decoder won't work for all base-64 and its variants. In fact even the encoder won't work in all contexts because some of them have mandatory line length limits. So ultimately the base64 encoder/decoder will need flags or needs to be some subset of all the variants; I think the current subset, after this pull, is a reasonable compromise but of course that's very subjective.