@svaarala 2017-10-30
The main optimization in the fast path (of the fast variant) is that it decodes two 4-character groups optimistically with only one error check branch for those total of 8 characters. So if the input is e.g. `Zm9vZm9v` it decodes to `foofoo` with one error check branch only. A trivial decoder has to check every character for validity (and the fallback of the fast variant does that of course) which involves a lot more branches and is consequently much slower.